Leave Ordinates.None in PostGisWriter to allow detection
PostGisWriter contains CheckOrdinates(), which looks at the given
geometry object to check which ordinates it contains; this is done
only if HandlOrdinates is set to None (i.e. the user hasn't
explicitly requested specific ordinates).

However, the HandleOrdinates property ORs XY into the given
value, so when the default constructor sets the property to None it
gets translated to XY. This means that CheckOrdinates() never gets
called.

This changes the HandleOrdinates property to only OR XY into the
given value if it is different from None.